Log 404 as INFO and put the url and reason to the title
Sentry groups all events by the title. So it is not possible to close or ignore just
some errors for specific urls but only all 404 error. Thats why the url has been added
to the title. Now every 404 should create a new error but not only an event for a 404.
As all of those informations are informations and not really warnings the level has been
adjusted to INFO.